bootcamp_meeting_ai_summaries/requirements.txt
fastapi
uvicorn
scikit-learn
motor
pymongo
python-dotenv
pydantic
streamlit
requests
pandas
matplotlib
python-multipart
google-generativeai
groq

